feat: abstract out multi stage output
